Corrected code after last checkins: username change,validate ownership,change royalty ticket, etc...
-fixed some warnings
-moved some boost-based tests from bitcoin_test to google test, changed to parameterized tests:
   ParseFixedPoint,  ParseInt32, Parse Int64, ParseDouble, ParseHex, HexStr
-changed ValidateOwnership,GetPastelIdAndTxIdWithTopHeightPerCopy to return optional tuple instead of string vector
-corrected some NFT->nft in rpc commands
-corrected NFT->nft in mn_tickets_validation.py,mn_expiration.py python tests
-fixes for mn python tests

+++ src/addrman.h
@@ -359,7 +359,7 @@ public:
             CAddrInfo &info = mapInfo[n];
             s >> info;
             mapAddr[info] = n;
-            info.nRandomPos = vRandom.size();
+            info.nRandomPos = static_cast<int>(vRandom.size());
             vRandom.push_back(n);
             if (nVersion != 1 || nUBuckets != ADDRMAN_NEW_BUCKET_COUNT) {
                 // In case the new table data cannot be used (nVersion unknown, or bucket count wrong),
